refactor(server): move domain logic into src/services
