+++ packages/emitsignal-server/src/http/messages/attachments.ts
@@ -2,14 +2,14 @@ import Elysia, { t } from 'elysia';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
-import { getUserPlan } from '#/lib/billing/get-user-plan';
-import { PLANS } from '#/lib/billing/plans';
-import { attachmentExpiresAt, messageRetentionDays } from '#/lib/billing/retention';
 import { prisma } from '#/lib/prisma';
 import { uploadAnonLimiter, uploadAuthLimiter } from '#/lib/rate-limit';
 import { FileStorageService } from '#/lib/storage';
 import { isAllowedMimeType } from '#/lib/storage/provider';
-import { resolveTopicCapabilities } from '#/lib/topic-access';
+import { getUserPlan } from '#/services/billing/get-user-plan';
+import { PLANS } from '#/services/billing/plans';
+import { attachmentExpiresAt, messageRetentionDays } from '#/services/billing/retention';
+import { resolveTopicCapabilities } from '#/services/topic-access';
 
 export const attachments = new Elysia({ prefix: '/messages' }).post(
     '/:id/attachments',
