+++ packages/emitsignal-server/src/http/subscriptions/messages.ts
@@ -3,7 +3,7 @@ import Elysia, { t } from 'elysia';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
 import { prisma } from '#/lib/prisma';
 import { readAnonLimiter, readAuthLimiter } from '#/lib/rate-limit';
-import { serializeMessage } from '#/lib/topic';
+import { serializeMessage } from '#/services/message';
 import { parseSubscriptionSettings } from '#/utils/subscription-settings';
 import { parseTagsQueryParam } from '#/utils/tags';
