+++ packages/emitsignal-server/src/http/subscriptions/unsubscribe.ts
@@ -2,7 +2,7 @@ import Elysia, { t } from 'elysia';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { prisma } from '#/lib/prisma';
-import { TOPIC_NAME_MAX_LENGTH } from '#/lib/topic';
+import { TOPIC_NAME_MAX_LENGTH } from '#/services/topic';
 
 export const unsubscribe = new Elysia({ prefix: '/subscriptions' }).delete(
     '/',
