+++ packages/emitsignal-server/src/http/topic/__tests__/publish-quota.spec.ts
@@ -21,9 +21,9 @@ mock.module('#/http/auth/plugin', () => ({
 }));
 
 import { publish } from '#/http/topic/publish';
-import { resetUserPlansForTests, setUserPlanForTests } from '#/lib/billing/get-user-plan';
-import { PLANS } from '#/lib/billing/plans';
-import { resetUsageForTests } from '#/lib/billing/usage';
+import { resetUserPlansForTests, setUserPlanForTests } from '#/services/billing/get-user-plan';
+import { PLANS } from '#/services/billing/plans';
+import { resetUsageForTests } from '#/services/billing/usage';
 
 describe('POST /topic/:name — plan quotas', () => {
     const app = new Elysia().use(publish);
