+++ packages/emitsignal-server/src/http/topic/get.ts
@@ -2,7 +2,7 @@ import Elysia from 'elysia';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { prisma } from '#/lib/prisma';
-import { resolveTopicCapabilities } from '#/lib/topic-access';
+import { resolveTopicCapabilities } from '#/services/topic-access';
 
 export const getTopic = new Elysia().get('/topics/:name', async ({ headers, params, status }) => {
     const topic = await prisma.topic.findUnique({
