+++ packages/emitsignal-server/src/http/topic/messages.ts
@@ -4,8 +4,8 @@ import { resolveUserId } from '#/http/auth/plugin';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
 import { prisma } from '#/lib/prisma';
 import { readAnonLimiter, readAuthLimiter } from '#/lib/rate-limit';
-import { serializeMessage } from '#/lib/topic';
-import { resolveTopicCapabilities } from '#/lib/topic-access';
+import { serializeMessage } from '#/services/message';
+import { resolveTopicCapabilities } from '#/services/topic-access';
 import { parseTagsQueryParam } from '#/utils/tags';
 
 export const messages = new Elysia().get(
