+++ packages/emitsignal-server/src/http/topic/publish.ts
@@ -2,15 +2,16 @@ import Elysia, { t } from 'elysia';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
-import { getUserLimits, getUserPlan } from '#/lib/billing/get-user-plan';
-import { messageExpiresAt, messageRetentionDays } from '#/lib/billing/retention';
-import { consumeDailyQuota, quotaExceededHeaders } from '#/lib/billing/usage';
 import { bus } from '#/lib/event-bus';
 import { prisma } from '#/lib/prisma';
 import { pushQueue, scheduleQueue } from '#/lib/queue';
 import { publishAnonLimiter, publishAuthLimiter } from '#/lib/rate-limit';
-import { getOrCreateTopic, serializeMessage, TopicNameError } from '#/lib/topic';
-import { resolveTopicCapabilities } from '#/lib/topic-access';
+import { getUserLimits, getUserPlan } from '#/services/billing/get-user-plan';
+import { messageExpiresAt, messageRetentionDays } from '#/services/billing/retention';
+import { consumeDailyQuota, quotaExceededHeaders } from '#/services/billing/usage';
+import { serializeMessage } from '#/services/message';
+import { getOrCreateTopic, TopicNameError } from '#/services/topic';
+import { resolveTopicCapabilities } from '#/services/topic-access';
 import { validateActions } from '#/utils/actions';
 import { duration } from '#/utils/duration';
 import { ANON_INLINE_MAX, validateMessageMedia } from '#/utils/media-refs';
