+++ packages/emitsignal-server/src/lib/__tests__/get-user-plan.spec.ts
@@ -10,8 +10,8 @@ import {
     invalidateUserPlanCache,
     resetUserPlansForTests,
     setUserPlanForTests,
-} from '#/lib/billing/get-user-plan';
-import { PLANS } from '#/lib/billing/plans';
+} from '#/services/billing/get-user-plan';
+import { PLANS } from '#/services/billing/plans';
 
 function subscriptionRow(overrides: Record<string, unknown> = {}) {
     return {
