+++ packages/emitsignal-server/src/lib/__tests__/usage.spec.ts
@@ -5,7 +5,7 @@ import {
     getDailyUsage,
     quotaExceededHeaders,
     resetUsageForTests,
-} from '#/lib/billing/usage';
+} from '#/services/billing/usage';
 
 beforeEach(() => {
     resetUsageForTests();
